+++ resources/views/admin/asset_library/photosphere_vr_view.blade.php
@@ -9,15 +9,11 @@
         <div class="col-md-8">
 
             <!-- a-frame //-->
-            <a-scene reset-camera embedded style="width:100%">
+            <a-scene embedded style="width:100%">
 
                 <a-assets>
-                    <img src="{{ asset('public/assets/admin/asset-library/images/loading.png') }}" id="loading">
+                    <img src="{{ asset('public/assets/admin/asset-library/images/loading.png') }}" id="loading" crossorigin="anonymous">
                 </a-assets>
-
-                <a-entity>
-                    <a-camera></a-camera>
-                </a-entity>
 
                 <a-sky id="default-sky" color="#000000"></a-sky>