+++ resources/views/admin/asset_library/video_vr_view.blade.php
@@ -9,16 +9,12 @@
         <div class="col-md-8">
 
             <!-- a-frame //-->
-            <a-scene reset-camera embedded style="width:100%">
+            <a-scene embedded style="width:100%">
 
                 <a-assets>
-                    <img src="{{ asset('public/assets/admin/asset-library/images/loading.png') }}" id="loading">
-                    <video id="video" src="{{ $uri }}">
+                    <img src="{{ asset('public/assets/admin/asset-library/images/loading.png') }}" id="loading" crossorigin="anonymous">
+                    <video id="video" src="{{ $uri }}" crossorigin="anonymous"></video>
                 </a-assets>
-
-                <a-entity>
-                    <a-camera></a-camera>
-                </a-entity>
 
                 <a-sky id="default-sky" color="#000000"></a-sky>