+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Configurable preview thumbnail images for content list
 - Dynamically configure GD or Imagick image driver depending on PHP configuration
 - Upload photo spheres: added checkbox to decide if original image dimension should be kept 
+- Upgraded to A-Frame 0.8.2
+- New theme added: Welcome theme
 
 ## 1.1.0